# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WhatsApp API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/chatrooms": {
            "get": {
                "tags": [
                    "Chatrooms"
                ],
                "summary": "List all chatrooms",
                "description": "Fetches a list of all chatrooms",
                "operationId": "ee4a9288947f684deb8e357e0e1eb56b",
                "responses": {
                    "200": {
                        "description": "List of chatrooms retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "General Chatroom"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "A place for general discussion"
                                                    },
                                                    "max_members": {
                                                        "type": "integer",
                                                        "example": 50
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-11-01T10:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Chatrooms"
                ],
                "summary": "Create a new chatroom",
                "description": "Creates a new chatroom with a given name and optional description",
                "operationId": "977e7e2bdc5de0f9e9c3f48ab33b37da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "General Chatroom"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A place for general discussion"
                                    },
                                    "max_members": {
                                        "type": "integer",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Chatroom created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Chatroom created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "General Chatroom"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "A place for general discussion"
                                                },
                                                "max_members": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-11-01T10:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/chatrooms/{chatroomId}/enter": {
            "post": {
                "tags": [
                    "Chatrooms"
                ],
                "summary": "Enter a chatroom",
                "description": "Allows a user to enter a chatroom",
                "operationId": "3afdd70b08bf23d2a542081087fe2e1d",
                "parameters": [
                    {
                        "name": "chatroomId",
                        "in": "path",
                        "description": "ID of the chatroom",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Entered the chatroom successfully"
                    },
                    "404": {
                        "description": "Chatroom not found"
                    },
                    "400": {
                        "description": "Chatroom is full or User is already in the chatroom"
                    }
                }
            }
        },
        "/api/chatrooms/{chatroomId}/leave": {
            "post": {
                "tags": [
                    "Chatrooms"
                ],
                "summary": "Leave a chatroom",
                "description": "Allows a user to leave a chatroom",
                "operationId": "232c4eafbe35ab93173d6057ef025ea6",
                "parameters": [
                    {
                        "name": "chatroomId",
                        "in": "path",
                        "description": "ID of the chatroom",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Left the chatroom successfully"
                    },
                    "404": {
                        "description": "Chatroom not found"
                    }
                }
            }
        },
...
    },
...
}